+++ src/main/java/com/global/api/entities/gpApi/GpApiAuthorizationRequestBuilder.java
@@ -34,6 +34,8 @@ public class GpApiAuthorizationRequestBuilder {
                 new JsonDoc()
                         .set("entry_mode", getEntryMode(builder, gateway.getGpApiConfig().getChannel())); // [MOTO, ECOM, IN_APP, CHIP, SWIPE, MANUAL, CONTACTLESS_CHIP, CONTACTLESS_SWIPE]
 
+        paymentMethod.set("narrative", !StringUtils.isNullOrEmpty(builder.getDynamicDescriptor()) ? builder.getDynamicDescriptor() : null);
+
         IPaymentMethod builderPaymentMethod = builder.getPaymentMethod();
         TransactionType builderTransactionType = builder.getTransactionType();
         TransactionModifier builderTransactionModifier = builder.getTransactionModifier();